requirements.txt
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
cattrs==24.1.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
Django==5.2.1
exceptiongroup==1.3.0
flatbuffers==25.2.10
h11==0.16.0
idna==3.10
jh2==5.0.9
niquests==3.14.1
numpy==2.2.6
openmeteo_requests==1.5.0
openmeteo_sdk==1.20.0
pandas==2.2.3
platformdirs==4.3.8
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qh3==1.5.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
urllib3-future==2.12.922
wassima==1.2.2
